add snapshot to cli and pow
